backend/requirements.txt
fastapi==0.112.0
SQLAlchemy==2.0.36
uvicorn==0.24.0
psycopg2-binary==2.9.10
python-dotenv
bcrypt==4.2.0
PyJWT==2.10.0